feat: Implemented Platform Specific Functionality
