+++ core/ui/src/androidInstrumentedTest/kotlin/com/google/samples/apps/nowinandroid/core/ui/NewsResourceCardTest.kt
