+++ core/ui/src/androidMain/AndroidManifest.xml
