+++ core/ui/src/commonMain/kotlin/com/google/samples/apps/nowinandroid/core/ui/FollowableTopicPreviewParameterProvider.kt
@@ -1,0 +1,68 @@
+/*
+ * Copyright 2023 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+@file:Suppress("ktlint:standard:max-line-length")
+
+package com.google.samples.apps.nowinandroid.core.ui
+
+import com.google.samples.apps.nowinandroid.core.model.data.FollowableTopic
+import com.google.samples.apps.nowinandroid.core.model.data.Topic
+import org.jetbrains.compose.ui.tooling.preview.PreviewParameterProvider
+
+/**
+ * This [PreviewParameterProvider](https://developer.android.com/reference/kotlin/androidx/compose/ui/tooling/preview/PreviewParameterProvider)
+ * provides list of [FollowableTopic] for Composable previews.
+ */
+class FollowableTopicPreviewParameterProvider : PreviewParameterProvider<List<FollowableTopic>> {
+    override val values: Sequence<List<FollowableTopic>>
+        get() = sequenceOf(
+            listOf(
+                FollowableTopic(
+                    topic = Topic(
+                        id = "2",
+                        name = "Headlines",
+                        shortDescription = "News we want everyone to see",
+                        longDescription = "Stay up to date with the latest events and announcements from Android!",
+                        imageUrl = "https://firebasestorage.googleapis.com/v0/b/now-in-android.appspot.com/o/img%2Fic_topic_Headlines.svg?alt=media&token=506faab0-617a-4668-9e63-4a2fb996603f",
+                        url = "",
+                    ),
+                    isFollowed = false,
+                ),
+                FollowableTopic(
+                    topic = Topic(
+                        id = "3",
+                        name = "UI",
+                        shortDescription = "Material Design, Navigation, Text, Paging, Accessibility (a11y), Internationalization (i18n), Localization (l10n), Animations, Large Screens, Widgets",
+                        longDescription = "Learn how to optimize your app's user interface - everything that users can see and interact with. Stay up to date on topics such as Material Design, Navigation, Text, Paging, Compose, Accessibility (a11y), Internationalization (i18n), Localization (l10n), Animations, Large Screens, Widgets, and many more!",
+                        imageUrl = "https://firebasestorage.googleapis.com/v0/b/now-in-android.appspot.com/o/img%2Fic_topic_UI.svg?alt=media&token=0ee1842b-12e8-435f-87ba-a5bb02c47594",
+                        url = "",
+                    ),
+                    isFollowed = true,
+                ),
+                FollowableTopic(
+                    topic = Topic(
+                        id = "4",
+                        name = "Testing",
+                        shortDescription = "CI, Espresso, TestLab, etc",
+                        longDescription = "Testing is an integral part of the app development process. By running tests against your app consistently, you can verify your app's correctness, functional behavior, and usability before you release it publicly. Stay up to date on the latest tricks in CI, Espresso, and Firebase TestLab.",
+                        imageUrl = "https://firebasestorage.googleapis.com/v0/b/now-in-android.appspot.com/o/img%2Fic_topic_Testing.svg?alt=media&token=a11533c4-7cc8-4b11-91a3-806158ebf428",
+                        url = "",
+                    ),
+                    isFollowed = false,
+                ),
+            ),
+        )
+}
