+++ core/ui/src/commonMain/kotlin/com/google/samples/apps/nowinandroid/core/ui/NewsFeed.kt
@@ -1,0 +1,148 @@
+/*
+ * Copyright 2022 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.samples.apps.nowinandroid.core.ui
+
+import androidx.annotation.ColorInt
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyListScope
+import androidx.compose.foundation.lazy.staggeredgrid.LazyStaggeredGridScope
+import androidx.compose.foundation.lazy.staggeredgrid.LazyVerticalStaggeredGrid
+import androidx.compose.foundation.lazy.staggeredgrid.StaggeredGridCells
+import androidx.compose.foundation.lazy.staggeredgrid.items
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.toArgb
+import androidx.compose.ui.unit.dp
+import com.google.samples.apps.nowinandroid.core.analytics.LocalAnalyticsHelper
+import com.google.samples.apps.nowinandroid.core.designsystem.theme.NiaTheme
+import com.google.samples.apps.nowinandroid.core.model.data.UserNewsResource
+import com.google.samples.apps.nowinandroid.core.ui.util.BrowserLauncher
+import org.jetbrains.compose.ui.tooling.preview.Preview
+import org.jetbrains.compose.ui.tooling.preview.PreviewParameter
+
+/**
+ * An extension on [LazyListScope] defining a feed with news resources.
+ * Depending on the [feedState], this might emit no items.
+ */
+fun LazyStaggeredGridScope.newsFeed(
+    feedState: NewsFeedUiState,
+    onNewsResourcesCheckedChanged: (String, Boolean) -> Unit,
+    onNewsResourceViewed: (String) -> Unit,
+    onTopicClick: (String) -> Unit,
+    onExpandedCardClick: () -> Unit = {},
+) {
+    when (feedState) {
+        NewsFeedUiState.Loading -> Unit
+        is NewsFeedUiState.Success -> {
+            items(
+                items = feedState.feed,
+                key = { it.id },
+                contentType = { "newsFeedItem" },
+            ) { userNewsResource ->
+                val analyticsHelper = LocalAnalyticsHelper.current
+                val backgroundColor = MaterialTheme.colorScheme.background.toArgb()
+
+                NewsResourceCardExpanded(
+                    userNewsResource = userNewsResource,
+                    isBookmarked = userNewsResource.isSaved,
+                    onClick = {
+                        onExpandedCardClick()
+                        analyticsHelper.logNewsResourceOpened(
+                            newsResourceId = userNewsResource.id,
+                        )
+                        launchCustomChromeTab(userNewsResource.url, backgroundColor)
+
+                        onNewsResourceViewed(userNewsResource.id)
+                    },
+                    hasBeenViewed = userNewsResource.hasBeenViewed,
+                    onToggleBookmark = {
+                        onNewsResourcesCheckedChanged(
+                            userNewsResource.id,
+                            !userNewsResource.isSaved,
+                        )
+                    },
+                    onTopicClick = onTopicClick,
+                    modifier = Modifier
+                        .padding(horizontal = 8.dp)
+                        .animateItem(),
+                )
+            }
+        }
+    }
+}
+
+fun launchCustomChromeTab(
+    uri: String,
+    @ColorInt toolbarColor: Int,
+) {
+    BrowserLauncher.openUrl(uri, toolbarColor)
+}
+
+/**
+ * A sealed hierarchy describing the state of the feed of news resources.
+ */
+sealed interface NewsFeedUiState {
+    /**
+     * The feed is still loading.
+     */
+    data object Loading : NewsFeedUiState
+
+    /**
+     * The feed is loaded with the given list of news resources.
+     */
+    data class Success(
+        /**
+         * The list of news resources contained in this feed.
+         */
+        val feed: List<UserNewsResource>,
+    ) : NewsFeedUiState
+}
+
+@Preview
+@Composable
+private fun NewsFeedLoadingPreview() {
+    NiaTheme {
+        LazyVerticalStaggeredGrid(columns = StaggeredGridCells.Adaptive(300.dp)) {
+            newsFeed(
+                feedState = NewsFeedUiState.Loading,
+                onNewsResourcesCheckedChanged = { _, _ -> },
+                onNewsResourceViewed = {},
+                onTopicClick = {},
+            )
+        }
+    }
+}
+
+@Preview
+@Preview
+@Composable
+private fun NewsFeedContentPreview(
+    @PreviewParameter(UserNewsResourcePreviewParameterProvider::class)
+    userNewsResources: List<UserNewsResource>,
+) {
+    NiaTheme {
+        LazyVerticalStaggeredGrid(columns = StaggeredGridCells.Adaptive(300.dp)) {
+            newsFeed(
+                feedState = NewsFeedUiState.Success(userNewsResources),
+                onNewsResourcesCheckedChanged = { _, _ -> },
+                onNewsResourceViewed = {},
+                onTopicClick = {},
+            )
+        }
+    }
+}
