+++ core/ui/src/nativeMain/kotlin/com/google/samples/apps/nowinandroid/core/ui/util/DragAndDropExt.native.kt
@@ -1,0 +1,40 @@
+/*
+ * Copyright 2025 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.samples.apps.nowinandroid.core.ui.util
+
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draganddrop.DragAndDropTransferData
+
+/**
+ * Platform-specific implementation of drag source modifier using official Compose API
+ */
+actual fun Modifier.dragSourceImpl(
+    content: String,
+    label: String,
+    onDragStart: (() -> Unit)?,
+    onDragEnd: (() -> Unit)?,
+): Modifier = Modifier
+
+/**
+ * Creates a drag and drop transfer data from content
+ */
+actual fun createDragData(
+    content: String,
+    label: String,
+): DragAndDropTransferData {
+    return DragAndDropTransferData()
+}
